03 editar

+++ tema-04/proyectos/02-articulos/02-Nuevo_Create/views/view.index.php
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="es">
 <head>
-    <?php include 'views/layouts/head.html'; ?>
+    <?php include 'views/layouts/layout.head.html'; ?>
     <title>Gestión de Artículos - Home </title>
 </head>
 <body>
@@ -33,30 +33,30 @@
                 </thead>
                 <tbody>
                     <!-- Mostramos cuerpo de la tabla -->
-                    <?php foreach ($tabla_articulos as $articulo): ?>
+                    <?php foreach ($array_articulos as $articulo): ?>
                         <tr>
                             <!-- Detalles de artículos -->
                             <td><?= $articulo->getId() ?></td>
                             <td><?= $articulo->getDescripcion() ?></td>
                             <td><?= $articulo->getModelo() ?></td>
                             <td><?= $marcas[$articulo->getMarca()] ?></td>
-                            <td><?= implode (', ', $articulos->mostrar_nombre_categorias($articulo->getCategorias())) ?></td>
+                            <td><?= implode(', ', $obj_tabla_articulos->mostrar_nombre_categorias($articulo->getCategorias())) ?></td>
                             <td class='text-end'><?= number_format($articulo->getUnidades(), 0, ',', '.') ?></td>
                             <td class='text-end'><?= number_format($articulo->getPrecio(), 2, ',', '.'). ' €' ?></td>
                             
                             <!-- Columna de acciones -->
                             <td>
-                                <a href="eliminar.php?id=<?$articulo->getId()?>" title="Eliminar"><i class="bi bi-trash-fill"></i></a>
-                                <a href="editar.php?id=<?$articulo->getId()?>" title="Editar"><i class="bi bi-pencil-square"></i></a>
-                                <a href="mostrar.php?id=<?$articulo->getId()?>" title="Mostrar"><i class="bi bi-eye-fill"></i></a>
+                                <div class="btn-group btn-group-toggle" data-toggle="buttons">
+                                <a href="eliminar.php?id=<?=$articulo->getId()?>" title="Eliminar" class="btn btn-danger" onclick="return confirm('Confirmar eliminación del artículo')"><i class="bi bi-trash-fill"></i></a>
+                                <a href="editar.php?id=<?=$articulo->getId()?>" title="Editar" class="btn btn-primary"><i class="bi bi-pencil-square"></i></a>
+                                <a href="mostrar.php?id=<?=$articulo->getId()?>" title="Mostrar" class="btn btn-warning"><i class="bi bi-eye-fill"></i></a>
+                                </div>
                             </td>
                         </tr>
                     <?php endforeach; ?>   
                 </tbody>
                 <tfoot>
-                    <tr>
-                        <td colspan="6">Nº Registros <?= count($tabla_articulos) ?></td>
-                    </tr>
+                    <tr><td colspan="6">Nº Registros <?= count($array_articulos) ?></td></tr>
                 </tfoot>
             </table>
         </div>
@@ -67,7 +67,7 @@
     <?php include 'views/partials/partial.footer.php';?>
 
     <!-- Bootstrap Javascript y popper -->
-    <?php include 'views/layouts/javascript.html';?>
+    <?php include 'views/layouts/layout.javascript.html';?>
     
  
 </body>
